Extract test cases in ParserTest into their own respective test files
